+++ ModelLibrary/Contract/DigitalOption/CashDigital.cs
@@ -8,18 +8,18 @@ namespace Contract.DigitalOption
 {
     public class CashDigital : IDigitalOption
     {
-        public double _strike { get; private set; }
-        public double _paymentValue { get; private set; }
-        public double _maturity{ get; private set; }
+        public decimal _strike { get; private set; }
+        public decimal _paymentValue { get; private set; }
+        public decimal _maturity{ get; private set; }
 
-        public CashDigital(double strike, double maturity, double paymentValue)
+        public CashDigital(decimal strike, decimal maturity, decimal paymentValue)
         {
             _strike = strike;
             _maturity = maturity;
             _paymentValue = paymentValue;
         }
 
-        public double payoff(double spot)
+        public decimal payoff(decimal spot)
         {
             if (spot > _strike)
                 return _paymentValue;
