+++ ModelLibrary/Contract/DigitalOption/IDigitalOption.cs
@@ -8,7 +8,7 @@ namespace Contract.DigitalOption
 {
     public interface IDigitalOption
     {
-        double payoff(double spot);
+        decimal payoff(decimal spot);
 
     }
 }
